remove MODEL_NAME from config and split training files to remove redundant code
